+++ src/domain/index.ts
@@ -1,0 +1,2 @@
+export * from "./services";
+export * from "./valueobjects";
