+++ src/domain/services/index.ts
@@ -1,1 +1,2 @@
 export * from "./insights-service";
+export * from "./svg-service";
