+++ src/domain/valueobjects/index.ts
@@ -1,0 +1,2 @@
+export * from "./Insights";
+export * from "./Options/Options";
