+++ src/presentation/index.ts
@@ -1,1 +1,2 @@
+export * from "./cli";
 export * from "./server";
