+++ tsconfig.json
@@ -12,9 +12,9 @@
     "baseUrl": ".",
     "paths": {
       "@assets/*": ["assets/*"],
-      "@util/*": ["src/util/*"],
-      "@domain/*": ["src/domain/*"],
-      "@presentation/*": ["src/presentation/*"]
+      "@util": ["src/util/index.ts"],
+      "@domain": ["src/domain/index.ts"],
+      "@presentation": ["src/presentation/index.ts"]
     }
   },
   "include": ["./src/**/*"],
